+++ models/roleAction.js
@@ -4,8 +4,8 @@ const sequelize = require("../db/sequelize");
 const RoleAction = sequelize.define(
     "RoleAction",
     {
-        roleId: { type: DataTypes.BIGINT, field: 'role_id', primaryKey: true },
-        actionId: { type: DataTypes.BIGINT, field: 'action_id', primaryKey: true },
+        roleId: { type: DataTypes.INTEGER, field: 'role_id', primaryKey: true },
+        actionId: { type: DataTypes.INTEGER, field: 'action_id', primaryKey: true },
     },
     {
         tableName: "role_actions",
